madameautomata-bot-python-autoreg/requirements.txt
discord.py>=2.4
python-dotenv>=1.0
SQLAlchemy>=2.0
asyncpg>=0.29
pydantic>=2.8
